Update binReassembly.sh
Identified strange behavior when using --trusted-contigs that result in short tandem repeats in the middle of contigs. I think this is from the use of trusted contigs in the construction of the assembly graph. I found that changing this to --untrusted-contigs removed this behavior. Also changed the use of "scaffolds.fasta" to "contigs.fasta" so that round N+1 is based on contigs, not scaffolds from Round N.